% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_colors.R
\docType{data}
\name{critcolors.map}
\alias{critcolors.map}
\title{Color Map Data}
\format{
An object of class \code{character} of length 8.
}
\usage{
critcolors.map
}
\description{
A data set containing the Hex codes for the original CritLab Colors.
}
\author{
Cecilia López-Martínez: \email{clm@crit-lab.org} /
}
\keyword{datasets}
