% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_colors.R
\name{critcolors}
\alias{critcolors}
\title{CritLab Color Palettes}
\usage{
critcolors(n, alpha = 1, begin = 0, end = 1, direction = 1)
}
\arguments{
\item{n}{The number of colors (\eqn{\ge 1}) to be in the palette.}

\item{alpha}{The alpha transparency, a number in \link{0,1}, see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in \link{0,1} at which the color map begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the color map ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are ordered from darkest to lightest. If -1, the order of colors is
reversed.}
}
\value{
\code{critcolors} returns a character vector, \code{cv}, of color hex
codes. This can be used either to create a user-defined color palette for
subsequent graphics by \code{palette(cv)}, a \code{col =} specification in
graphics functions or in \code{par}.
}
\description{
This function creates a vector of \code{n} equally spaced colors
along the critcolors color map.
}
\author{
Cecilia López-Martínez: \email{clm@crit-lab.org} /
Code adapted from the "viridis" package by Simon Garnier: \email{garnier@njit.edu} / \href{https://twitter.com/sjmgarnier}{@sjmgarnier}
}
