% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_colors.R
\name{critcolors_pal}
\alias{critcolors_pal}
\title{Critcolors Color Palettes}
\usage{
critcolors_pal(alpha = 1, begin = 0, end = 1, direction = 1)
}
\arguments{
\item{alpha}{The alpha transparency, a number in \link{0,1}, see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in \link{0,1} at which the color map begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the color map ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are ordered from darkest to lightest. If -1, the order of colors is
reversed.}
}
\description{
A wrapper function around critcolors to
turn it into a palette function compatible with
\code{\link[ggplot2]{discrete_scale}}.
}
\author{
Cecilia López-Martínez: \email{clm@crit-lab.org} /
Code adapted from the "viridis" package by

Bob Rudis: \email{bob@rud.is} / \href{https://twitter.com/hrbrmstr}{@hrbrmstr}

Simon Garnier: \email{garnier@njit.edu} / \href{https://twitter.com/sjmgarnier}{@sjmgarnier}
}
