% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_colors.R
\name{scale_fill_critcolors}
\alias{scale_fill_critcolors}
\title{CritColors Color Scales for ggplot2}
\usage{
scale_fill_critcolors(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE
)
}
\arguments{
\item{...}{Parameters to \code{\link[ggplot2]{discrete_scale}} if
\code{discrete == TRUE}, or \code{\link[ggplot2]{scale_fill_gradientn}}/
\code{\link[ggplot2]{scale_color_gradientn}} if \code{discrete == FALSE}.}

\item{alpha}{The alpha transparency, a number in \link{0,1}, see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in \link{0,1} at which the color map begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the color map ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are as output by \code{\link[viridis]{viridis_pal}}. If -1, the order
of colors is reversed.}

\item{discrete}{Generate a discrete palette? (default: \code{FALSE} -
generate continuous palette).}
}
\description{
Scale functions (fill and colour/color) for
\code{\link[ggplot2]{ggplot2}}.
}
\author{
Cecilia López-Martínez: \email{clm@crit-lab.org} /
Code adapted from the "viridis" package by

Noam Ross \email{noam.ross@gmail.com} / \href{https://twitter.com/noamross}{@noamross}

Bob Rudis \email{bob@rud.is} / \href{https://twitter.com/hrbrmstr}{@hrbrmstr}

Simon Garnier: \email{garnier@njit.edu} / \href{https://twitter.com/sjmgarnier}{@sjmgarnier}
}
